ButtonLineEdit: Move into its own file under `rare/widgets`
